requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
audioread==2.1.9
beautifulsoup4==4.11.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
decorator==5.1.1
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.46.0
idna==3.3
importlib-metadata==4.11.3
joblib==1.1.0
kiwisolver==1.4.2
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.7
matplotlib==3.5.1
multidict==6.0.2
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
pooch==1.6.0
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.2
pydub==0.25.1
pyparsing==3.0.8
python-dateutil==2.8.2
pytorch-lightning==1.6.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.8.2
torchvision==0.12.0
tqdm==4.64.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wget==3.2
yarl==1.7.2
zipp==3.8.0
